+++ modules/hanna-react/src/SiteSearchCurtain.tsx
@@ -1,40 +1,38 @@
-import React, { FocusEvent, ReactNode, useState } from 'react';
-import { useOnMount } from '@hugsmidjan/react/hooks';
+import React, { FocusEvent, ReactNode, useEffect } from 'react';
+import { useLaggedState } from '@hugsmidjan/react/hooks';
 import getBemClass from '@hugsmidjan/react/utils/getBemClass';
 
 import { isPreact } from './utils/env';
 
-export type SiteSearchCurtain = {
+export type SiteSearchCurtainProps = {
   children: ReactNode;
 };
 
-export const SiteSearchCurtain = (props: SiteSearchCurtain) => {
-  const [focused, setFocused] = useState(false);
-  let blurTimeout: NodeJS.Timeout | undefined;
+export const SiteSearchCurtain = (props: SiteSearchCurtainProps) => {
+  const [focused, setFocused] = useLaggedState(false);
 
-  useOnMount(() => {
-    const closeSearch = (e: KeyboardEvent) => {
-      if (e.key === 'Escape') {
-        setFocused(false);
-      }
-    };
-    document.addEventListener('keyup', closeSearch);
+  useEffect(
+    () => {
+      const closeSearch = (e: KeyboardEvent) => {
+        if (e.key === 'Escape') {
+          setFocused(false);
+        }
+      };
+      document.addEventListener('keyup', closeSearch);
 
-    return () => {
-      document.removeEventListener('keyup', closeSearch);
-    };
-  });
+      return () => document.removeEventListener('keyup', closeSearch);
+    },
+
+    // eslint-disable-next-line react-hooks/exhaustive-deps
+    []
+  );
 
   const focusHandler = () => {
-    blurTimeout && clearTimeout(blurTimeout);
     setFocused(true);
   };
   const blurHandler = (e: FocusEvent) => {
     if (!e.relatedTarget || !e.currentTarget.contains(e.relatedTarget as Node)) {
-      blurTimeout = setTimeout(() => {
-        setFocused(false);
-        blurTimeout = undefined;
-      }, 100);
+      setFocused(false, 100);
     }
   };